+++ pkg/servicemodel/registry/registery.go
@@ -7,25 +7,26 @@ package registry
 import (
 	"sync"
 
+	"github.com/onosproject/ran-simulator/pkg/servicemodel"
+
 	"github.com/onosproject/onos-lib-go/pkg/errors"
-	"github.com/onosproject/ran-simulator/pkg/servicemodels"
 )
 
 // ServiceModelRegistry stores list of registered service models
 type ServiceModelRegistry struct {
 	mu            sync.Mutex
-	serviceModels map[ID]servicemodels.ServiceModel
+	serviceModels map[ID]servicemodel.ServiceModel
 }
 
 // NewServiceModelRegistry creates a service model registry
 func NewServiceModelRegistry() *ServiceModelRegistry {
 	return &ServiceModelRegistry{
-		serviceModels: make(map[ID]servicemodels.ServiceModel),
+		serviceModels: make(map[ID]servicemodel.ServiceModel),
 	}
 }
 
 // RegisterServiceModel registers a service model
-func (s *ServiceModelRegistry) RegisterServiceModel(id ID, sm servicemodels.ServiceModel) error {
+func (s *ServiceModelRegistry) RegisterServiceModel(id ID, sm servicemodel.ServiceModel) error {
 	if _, exists := s.serviceModels[id]; exists {
 		return errors.New(errors.AlreadyExists, "the service model already registered")
 	}
