--- pkg/servicemodels/servicemodel.go
@@ -1,16 +1,0 @@
-// SPDX-FileCopyrightText: 2020-present Open Networking Foundation <info@opennetworking.org>
-//
-// SPDX-License-Identifier: LicenseRef-ONF-Member-1.0
-
-package servicemodels
-
-import (
-	"github.com/onosproject/onos-e2t/api/e2ap/v1beta1/e2appducontents"
-)
-
-// ServiceModel service model interface
-type ServiceModel interface {
-	ProcessSubscription(request *e2appducontents.RicsubscriptionRequest) (response *e2appducontents.RicsubscriptionResponse, failure *e2appducontents.RicsubscriptionFailure, err error)
-	ProcessSubscriptionDelete(request *e2appducontents.RicsubscriptionDeleteRequest) (response *e2appducontents.RicsubscriptionDeleteResponse, failure *e2appducontents.RicsubscriptionDeleteFailure, err error)
-	ProcessControl(request *e2appducontents.RiccontrolRequest) (response *e2appducontents.RiccontrolAcknowledge, failure *e2appducontents.RiccontrolFailure, err error)
-}
